Rename old cli
