+++ connectors/service_cli.py
